finished my work on datatables on develop branch,working properly

+++ src/main/java/com/softserve/edu/dao/RequestDao.java
@@ -5,7 +5,7 @@ import java.util.List;
 import com.softserve.edu.entity.Request;
 import com.softserve.edu.entity.exceptions.RequestAlreadySentException;
 
-public interface RequestDao extends GenericDao<Request, Long> {
+public interface RequestDao extends GenericDao<Request, Integer> {
 	public boolean checkRequest(Request request) throws RequestAlreadySentException;
 	
 	public List<Request> getAllApprovedRequestsByHostingId(int hostingId);
